Closes #106
Create a DefaultVendorDTO to be used when the request body does not have the properties like "address" and "url".
This DTO is used as the default implementation for deserializing the vendorDTO.
Thus, when patching only the name of the vendor, the DefaultVendorDTO is used. Since the service will search the repository with the id passed, it will recover the correct vendor and update the name.